Clarify how to show the users organization
I looked _several times_ to find how to add the organization name or ID to the JWT. but kept overlooking this.

The `urn:zitadel:iam:user:resourceowner` claim adds the users' organization. But because the word organization was missing, it was very much non-obvious. 

This fix proposes a clarification of the description to clarify this.