feat(actions): migrate to v2beta API (#9489)
# Which Problems Are Solved

The current API esp. the versioning led to a lot of confusion since
there is/was a V1 marked as General available with multiple services, V2
marked the same and additionally a V3 in preview.
Since user v3 and the corresponding user schema services have been
postponed, the remaining WebKey service and Action service are the only
remaining services in the v3 API.
As these features matured, we want to bring them into a more stable beta
phase before moving them into GA later on.
This PR therefore migrates the service to V2 Beta to simplify the
versioning.

All V3 Alpha endpoints are removed. Please check out the changes listed
below if you already implemented the current API.

# How the Problems Are Solved

- Migrated from v3alpha to v2beta
- `zitadel.resources.action.v3alpha.ZITADELActions` ->
`zitadel.action.v2beta.ActionService`
  - `/resources/v3alpha/actions` -> `/v2beta/actions`
- added v2 endpoints to the API reference documentation and removed v3
endpoints
  - updated action guides in the docs accordingly
- Additional changes to the endpoints
  - the instance cannot be provided by the request anymore
  - Creation
    - moved configuration into request (from separate message)
- response provides id and `creation_date` (instead of details message)
  - Patch
    - moved to Update
    - moved configuration into request (from separate message)
    - response provided `change_date` (instead of details message)
  - Deletion
- successful deletion of an (existing) target will return a
`deletion_date` (instead of details message)
    - not existing or already removed target returns no error anymore
  - List
- Type changed from `GetTarget` to `Target` and `GetExecution` to
`Execution`
- moved details (id, creation_date, change_date) into the `Execution`
object from a dedicated message
  - Pagination
    - added pagination messages to reuse 
  - Filter
    - added filter messages to reuse

# Additional Changes

Replace this example text with a concise list of additional changes that
this PR introduces, that are not directly solving the initial problem
but are related.
For example:
- The docs explicitly describe that the property XY is mandatory

# Additional Context

Closes #9410

---------

Co-authored-by: Livio Spring <livio.a@gmail.com>