feat: Web Keys management (#9526)
# Which Problems Are Solved
Adds Web Keys managment to the Instance settings.

# How the Problems Are Solved
Added new page to the Instance settings.

![grafik](https://github.com/user-attachments/assets/c66bc8d7-b277-453c-9f5c-d9629cb49cb2)

# Additional Changes
Removed dataSize input from refresh table as it's not actually used
anywhere.

# Additional Context
It should always be **Web Keys** not Webkeys etc. as that is the
official naming, if I missed it somewhere please mark it.
At the moment the code won't compile because the @zitadel/client package
is not released with the new functionality.
In the code at someplaces the V3 Alpha proto is referenced because I
wasn't able to npm link @zitadel/client and @zitadel/proto at the same
time. I will fix this as soon as we have released the @zitadel/client.
Translations are also missing at the moment as I'd like @skewis6 to take
a look at the english texts first.

- Closes #8033

Open Todo's
- [ ] Fix V3 Alpha importants once @zitadel/client is released
- [ ] Add Translations
- [ ] Fix warn dialog with correct translations aswell
- [ ] Remove all todo's
- [ ] Merge main into this branch

---------

Co-authored-by: Livio Spring <livio.a@gmail.com>
Co-authored-by: Elio Bischof <elio@zitadel.com>
Co-authored-by: Kenta Yamaguchi <56732734+KEY60228@users.noreply.github.com>
Co-authored-by: Harsha Reddy <harsha.reddy@klaviyo.com>
Co-authored-by: Silvan <27845747+adlerhurst@users.noreply.github.com>
Co-authored-by: Livio Spring <livio@zitadel.com>
Co-authored-by: Stefan Benz <46600784+stebenz@users.noreply.github.com>
Co-authored-by: Max Peintner <max@caos.ch>