fix(authz): simplify system permission mapping (#9627)
# Which Problems Are Solved

Remove the need of maintaining a channel for System user permission
mapping, while still ensuring the mapping is built only once per
request.

# How the Problems Are Solved

Build the slice of `SystemUserPermissionsDBQuery` once, when the context
is being created in the context middleware. This slice will remain valid
during the request lifetime as part of the context. This removes the
need of "caching" and additional synchronization.

# Additional Changes

```sh
 gci write . --skip-generated -s standard -s default -s 'prefix(github.com/zitadel/zitadel)' --custom-order
```

# Additional Context

- Follow up on discussion
https://github.com/zitadel/zitadel/pull/9460#discussion_r2006769497