fix(console): add loginV2 feature flag (#9682)
# Which Problems Are Solved

While there was already an option to enable the new login UI for a
specific application, there was no possibility to do so for the whole
instance yet. This prevents an easy switch to the new hosted UI or when
running a self hosted login UI.

# How the Problems Are Solved

Added the LoginV2 settings to the feature toggle screen.

# Additional Changes

Simplified the feature flags code a bit and made typings more robust.

# Additional Context

Part of https://github.com/zitadel/zitadel/issues/9481