chore(deps): group easily mergable dependabot PRs (#9689)
# Which Problems Are Solved

Dependabot creates a PR for each individual dependency update. This is
noisy and toil.

# How the Problems Are Solved

The easily mergable updates are grouped into fewer PRs.
Major updates and security updates are not grouped, as they might need
special attention.
For the github-actions updates, also the major updates are grouped, as
we have immediate and good enough feedback from the pipeline if a new
major doesn't work.

# Additional Changes

The e2e dependencies are also auto-updated.

# Additional Context

- We aggreed on grouping the dependabot PRs in todays Scrum daily.
- I don't know how to test the config. We might just have to merge it
and look how it goes :man_shrugging: