docs: add note about required version to zitadel v3 feature documentation (#9683)
# Which Problems Are Solved

Since were about to merge and publish a first RC of Zitadel V3, there
are some inconsistencies to v2 in the documentation.
Also since zitadel.cloud will not be running v3 yet, there might be some
confusion.
To prevent as much as possible, we want at least be able to point out
which version is required.

# How the Problems Are Solved

Added a note on the actions v2 and webkeys documentations that it
requires Zitadel v3, resp. that it documents the state of that version.
Added the same note to the hosted login v2 description, since some
features are only supported in v3.

# Additional Changes

None

# Additional Context

Relates to the upcomming merge of v3 to main / release of the first RC.