feat(console): add Actions V2 UI (#9591)
# Which Problems Are Solved
This PR add's the the creation and management of Actions V2 to the
Console.

# How the Problems Are Solved
In the Instance setting's under the Category Actions the Actions and
Targets can be configured and created.


![grafik](https://github.com/user-attachments/assets/9b78e737-afae-487b-8033-9139a184e830)

![grafik](https://github.com/user-attachments/assets/2de144ca-cec7-4795-9b03-47437b50f85d)

# Additional Changes
Some work went into our settings list to make it more generic and in the
future easier to extend and use the router instead of custom conditional
rendering.

# Additional Context
- Closes #7248 

Some UX and additional features will be added in the future.

---------

Co-authored-by: Max Peintner <peintnerm@gmail.com>
Co-authored-by: Livio Spring <livio.a@gmail.com>