feat: manage federated logout option on SAML IdP (#9841)
# Which Problems Are Solved

As part of federated logout
([#9228](https://github.com/zitadel/zitadel/issues/9228)),
administrators need to be able to manage the federated logout option of
SAML IdPs.

# How the Problems Are Solved

This PR adds the option to manage the option of the V1 APIs (admin and
management) as well as retrieve it on the V2 IDP API. Additionally,
Console is updated to provide the possibility as well.

# Additional Changes

None

# Additional Context

part of #9228

---------

Co-authored-by: Silvan <27845747+adlerhurst@users.noreply.github.com>
Co-authored-by: Zach Hirschtritt <zachary.hirschtritt@klaviyo.com>