feat: centralize id and ts filters (#9969)
# Which Problems Are Solved

Creating a separate message for each id and timestamp filter property is
cumbersome, not easy to use and prone to inconsistencies accross the
API. We should be able to reuse IDFilter and TimestampFilter in
https://github.com/zitadel/zitadel/pull/9338/files and other new
resource APIs. This also speeds up the implementation of
https://github.com/zitadel/zitadel/issues/6305.

# How the Problems Are Solved

The IDFilter and TimestampFilter messages are moved from the user
package to a central filter package. The v2beta ProjectService and the
v2 UserService are changed so they use the central messages.

# Additional Context

- The idea came up during updating
https://github.com/zitadel/zitadel/pull/9338/files
- Accelerates https://github.com/zitadel/zitadel/issues/6305