fix(queue): reset projection list before each `Register` call (#10001)
if Zitadel was started using `start-from-init` or `start-from-setup`
there were rare cases where a panic occured when
`Notifications.LegacyEnabled` was set to false. The cause was a list
which was not reset before refilling.

The list is now reset before each time it gets filled.

Ensure all contexts are canceled for the init and setup functions for
`start-from-init- or `start-from-setup` commands.

none