feat: implementation internal permission service for administrators (#10081)
# Which Problems Are Solved

The commands for the resource based v2beta InteralPermissionService API
are added.
Administrators, previously knows as memberships, give a user in a
specific organization and project context roles.
The project can be owned or granted.
The give roles give the user permissions to manage different resources
in Zitadel.

# How the Problems Are Solved

### New Methods

- CreateAdministrator
- UpdateAdministrator
- DeleteAdministrator
- ListAdministrators

# Additional Changes

- correct existence checks for resources, for example you can only be an
administrator on an existing project
- combined all member tables to singular query for the administrators
- add permission checks for command an query side functionality
- combined functions on command side where necessary for easier
maintainability

# Additional Context

Contributes to #9165